Refactor: move channel mux to util
